requirements.txt
pyyaml
numpy
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
flask
transformers